Remove unnecessary indexes
